Shortened/Simplified code. Clarified comments.
Removed the option to specify filename as a command line argument. The filename can be changed in the program if this is truly desired. 

Switched to loading the entire file into a list rather than rereading it line by line for every token. Seriously, was I drunk when I originally wrote this?

Removed some unnecessary code that was a holdover from utilities.